Refactor `GenerateRollupConfig` and `andTxSucceeded` to common testing package to reduce duplicated lines of code.
