+++ protocol/sol-implementation/BUILD.bazel
@@ -43,6 +43,7 @@ go_test(
         "//solgen/go/ospgen",
         "//solgen/go/rollupgen",
         "//state-manager",
+        "//testing",
         "//testing/mocks",
         "//testing/setup:setup_lib",
         "//util",
@@ -50,7 +51,6 @@ go_test(
         "@com_github_ethereum_go_ethereum//accounts/abi/bind/backends",
         "@com_github_ethereum_go_ethereum//common",
         "@com_github_ethereum_go_ethereum//core",
-        "@com_github_ethereum_go_ethereum//core/types",
         "@com_github_ethereum_go_ethereum//crypto",
         "@com_github_offchainlabs_nitro//util/headerreader",
         "@com_github_stretchr_testify//require",
