+++ testing/mocks/BUILD.bazel
@@ -2,6 +2,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "mocks",
+    testonly = 1,
     srcs = ["mocks.go"],
     importpath = "github.com/OffchainLabs/challenge-protocol-v2/testing/mocks",
     visibility = ["//visibility:public"],
