+++ testing/rollup_config.go
@@ -1,0 +1,43 @@
+package challenge_testing
+
+import (
+	"math/big"
+
+	"github.com/OffchainLabs/challenge-protocol-v2/solgen/go/rollupgen"
+	"github.com/ethereum/go-ethereum/common"
+)
+
+func GenerateRollupConfig(
+	prod bool,
+	wasmModuleRoot common.Hash,
+	rollupOwner common.Address,
+	chainId *big.Int,
+	loserStakeEscrow common.Address,
+	challengePeriodSeconds *big.Int,
+	miniStakeValue *big.Int,
+) rollupgen.Config {
+	var confirmPeriod uint64
+	if prod {
+		confirmPeriod = 45818
+	} else {
+		confirmPeriod = 20
+	}
+	return rollupgen.Config{
+		ChallengePeriodSeconds:   challengePeriodSeconds,
+		MiniStakeValue:           miniStakeValue,
+		ConfirmPeriodBlocks:      confirmPeriod,
+		ExtraChallengeTimeBlocks: 200,
+		StakeToken:               common.Address{},
+		BaseStake:                big.NewInt(100),
+		WasmModuleRoot:           wasmModuleRoot,
+		Owner:                    rollupOwner,
+		LoserStakeEscrow:         loserStakeEscrow,
+		ChainId:                  chainId,
+		SequencerInboxMaxTimeVariation: rollupgen.ISequencerInboxMaxTimeVariation{
+			DelayBlocks:   big.NewInt(60 * 60 * 24 / 15),
+			FutureBlocks:  big.NewInt(12),
+			DelaySeconds:  big.NewInt(60 * 60 * 24),
+			FutureSeconds: big.NewInt(60 * 60),
+		},
+	}
+}
